% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtranscribe-package.R
\docType{package}
\name{rtranscribe-package}
\alias{rtranscribe}
\alias{rtranscribe-package}
\title{rtranscribe: Speech Recognition with 'transcribe.cpp'}
\description{
Bindings to 'transcribe.cpp', a C++ automatic speech recognition library supporting many model families (Whisper, Parakeet, Moonshine, Voxtral and others) in the 'GGUF' format. Transcribe audio to text with segment-, word- and token-level timestamps, speaker diarization, batch processing and incremental streaming. The 'transcribe.cpp' and 'ggml' sources are bundled and built from source, so no separate installation of the library is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JBGruber/rtranscribe}
  \item Report bugs at \url{https://github.com/JBGruber/rtranscribe/issues}
}

}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{johannes.gruber@gesis.org} [copyright holder]

Authors:
\itemize{
  \item Johannes B. Gruber \email{johannes.gruber@gesis.org} [copyright holder]
}

Other contributors:
\itemize{
  \item The transcribe.cpp authors (Bundled 'transcribe.cpp' sources) [copyright holder]
  \item The ggml authors (Bundled 'ggml' sources) [copyright holder]
  \item Rich Geldreich and Tenacious Software LLC (Bundled 'miniz' sources) [copyright holder]
  \item Mozilla Foundation (Bundled 'llamafile' tinyBLAS sources) [copyright holder]
}

}
\keyword{internal}
