% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{transcribe_audio_duration}
\alias{transcribe_audio_duration}
\title{Audio duration in seconds}
\usage{
transcribe_audio_duration(x, sample_rate = 16000)
}
\arguments{
\item{x}{A numeric PCM vector or a path to an audio file.}

\item{sample_rate}{Sample rate of \code{x} when it is a numeric vector.}
}
\value{
A single number.
}
\description{
Audio duration in seconds
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
transcribe_audio_duration(pcm)
\dontshow{\}) # examplesIf}
}
