% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{transcribe_version}
\alias{transcribe_version}
\title{Library version}
\usage{
transcribe_version()
}
\value{
A list with the bundled transcribe.cpp \code{version} and the git
\code{commit} it was built from.
}
\description{
Library version
}
\examples{
transcribe_version()

}
